+++ changes.md
@@ -1,5 +1,5 @@
 #### 2.0.0
-* WIP
+* Moved "Advanced Settings" to the "Help" tab out of the "Developer" tab.
 * Added option to redirect certain property types to parent on front-end.
 * Changed the WPP settings export to use JSON extension.
 * Fixed property overview filter to work when pressing enter.